Restore repository to normal
I was in a subfolder when I tried to add the index.html and styles.css
files to the repository on git, by using the command git add
../index.html etc... this created copies of the files in the parent
directory which greatly confused me when I realized there were suddenly
two copies of the same file in the parent directory. After realizing
this, I deleted them thinking they were just copies. However, they were
actually the most up to date files that I had just committed and the
ones in the css-methods file were the old ones. This required me to
go back into the old files and remake the same changes.